Fix pagination
With the release of https://github.com/ponylang/http/pull/65, all headers are
stored converted to lowercase. Although `Payload.apply` looks up headers in a
case-insensitive manner, we copy the headers as-is into `Response`[0], so we
lose that feature. The fix then is to look all headers converted to lowercase.

[0]: https://github.com/ponylang/pony-sync-helper/blob/b0a63019cb809f9ac883fe678531eaade8a79d8a/asking/asking.pony#L86-L88