Use sync helper to post "good first issues" every week
This commit updates the sync helper to take a new required command
line parameter "label". It then uses this new ability in an additional
workflow to post all ponylang organizations "Good first issues" to the
Zulip "contribute to Pony" stream once a week.

This was discussed at a sync in the not so distant past as a desirable
thing to do as it will increase visibility of easy to get started with
issues and we hope, result in more people contributing.