Fix truncation of good first issues when posting to Zulip
The Zulip API has a max message size and our "good first issues"
posting hits it resulting in message truncation and a loss of
content.

This commit adds a Python script to batch up "good first issues"
content into 1 or more messages so that it doesn't get truncated.

Also includes new image to run in as we need additional dependencies
that the "post sync" image doesn't need.

And this adds a PR check to link the python script and make sure it
is all kosher.

Fixes #18